# E-commerce2/Controllers/Admin/ProductsController.cs
using ECommerce2.DTOs;
using ECommerce2.Services.Interfaces;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace ECommerce2.Controllers.Admin
{
    [ApiController]
    [Route("api/admin/[controller]")]
    [Authorize(Roles = "Admin")]
    public class ProductsController : ControllerBase
    {
        private readonly IProductService _productService;

        public ProductsController(IProductService productService)
        {
            _productService = productService;
        }

        [HttpPost]
        public async Task<IActionResult> Create([FromBody] CreateProductDto dto)
        {
            var result = await _productService.CreateAsync(dto);
            return result.IsSuccess
                ? CreatedAtAction(nameof(GetById), new { id = result.Value }, result.Value)
                : BadRequest(new { error = result.Error });
        }

        [HttpGet("{id:int}")]
        public async Task<IActionResult> GetById(int id)
        {
            var product = await _productService.GetByIdAsync(id);
            return product is null ? NotFound() : Ok(product);
        }

        [HttpGet]
        public async Task<IActionResult> GetProducts([FromQuery] ProductQueryParameters parameters)
        {
            var products = await _productService.GetAdminProductsAsync(parameters);
            return Ok(products);
        }

        [HttpPut("{id:int}")]
        public async Task<IActionResult> UpdateProduct(int id, [FromBody] UpdateProductDto dto)
        {
            var result = await _productService.UpdateAsync(id, dto);
            if (!result.IsSuccess)
                return BadRequest(new { error = result.Error });

            return Ok();
        }

        [HttpDelete("{id:int}")]
        public async Task<IActionResult> DeleteProduct(int id)
        {
            var result = await _productService.DeleteAsync(id);
            if (!result.IsSuccess)
                return BadRequest(new { error = result.Error });

            return Ok();
        }

        public record UpdateStatusRequest(bool Status);

        [HttpPatch("{id:int}/status")]
        public async Task<IActionResult> UpdateProductStatus(int id, [FromBody] UpdateStatusRequest request)
        {
            var result = await _productService.UpdateStatusAsync(id, request.Status);
            if (!result.IsSuccess)
                return BadRequest(new { error = result.Error });

            return Ok();
        }
    }
}